Move Register traits into register
