+++ crates/cpu/src/futures/getat.rs
@@ -1,5 +1,5 @@
 use super::Reader;
-use crate::area::Bits16;
+use crate::registers::Bits16;
 use crate::{RegisterBus, Registers};
 use memory::{Async, Memory};
 use shared::Error;
