+++ crates/cpu/src/lib.rs
@@ -1,18 +1,11 @@
-pub mod area;
-pub(crate) mod arithmetic;
 pub mod bus;
 pub mod cpu;
 pub(crate) mod executor;
-mod flags;
 pub mod futures;
 pub mod interface;
-pub(crate) mod logical;
 mod opcodes;
 pub mod registers;
-pub(crate) mod shift;
 
-pub(crate) use crate::arithmetic::Arithmetic;
 pub use crate::bus::RegisterBus;
 pub use crate::cpu::Cpu;
-pub use crate::interface::{NewRegisters, Registers};
-pub(crate) use flags::Flags;
+pub use crate::interface::{New, Registers};
