+++ crates/cpu/src/opcodes/logical.rs
@@ -1,7 +1,6 @@
-use crate::area::Bits8;
+use crate::registers::{Bits8, Logical};
 use crate::cpu::Registers;
 use crate::futures::{LogicalHL, LogicalNext, Operation};
-use crate::logical::Logical;
 use memory::Memory;
 use num_enum::TryFromPrimitive;
 
@@ -170,7 +169,7 @@ impl Logic {
 #[cfg(test)]
 mod test_logic_opcodes {
     use super::Logic;
-    use crate::area::{Bits16, Bits8, Flag};
+    use crate::registers::{Bits16, Bits8, Flag};
     use crate::{executor, RegisterBus, Registers};
     use memory::Memory;
