+++ crates/cpu/src/opcodes/src.rs
@@ -1,4 +1,4 @@
-use crate::area::{Bits16, Bits8};
+use crate::registers::{Bits16, Bits8};
 use crate::bus::RegisterBus;
 use crate::cpu::Registers;
 use crate::futures::{GetAt, SetAt};
