+++ crates/cpu/src/registers/arithmetic.rs
